pyproject.toml
[project]
name = "zarr_creator"
version = "0.1.0"
description = "zarr creation from harmonie dini forecasts"
authors = [
    {name = "Kasper Stener Hintz", email = "kah@dmi.dk"},
    {name = "Eleni Briola", email = "elb@dmi.dk"},
]
dependencies = [
    "loguru>=0.7.3",
    "rechunker>=0.5.2",
    "fsspec>=2024.12.0",
    "dmidc[harmonie-grib] @ file:///home/ec2-user/data-catalog",
    "ipdb>=0.13.13",
    "dask<=2024.11.2",
    "zarrdump>=0.4.2",
    "s3fs>=2025.2.0",
    "cf-xarray>=0.10.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "nbconvert>=7.16.6",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
